refactor(app): move create mutations with ID/timestamp logic to features
useCreatePackMutation, useAddPackItemMutation, and useCreateTripMutation
move to features/ since they contain client-side logic (UUID generation,
timestamp injection, value defaulting). Plain pass-through mutations
stay in entity queries.ts.
