+++ packages/app/src/entities/pack/index.ts
@@ -1,6 +1,4 @@
 export {
-  useAddPackItemMutation,
-  useCreatePackMutation,
   useDeletePackItemMutation,
   useDeletePackMutation,
   usePack,
