+++ packages/app/src/entities/trip/index.ts
@@ -1,5 +1,4 @@
 export {
-  useCreateTripMutation,
   useDeleteTripMutation,
   useTrip,
   useTrips,
